refactor(shell): split main.cpp into focused modules
Split the 1737-line main.cpp into 11 files. AuraShellWindow moved from
anonymous namespace to header-declared class with methods split across
3 .cpp files (window core, panels, refresh). Extracted size_metrics,
shell_utils, and stylesheet_builder into aura_shell_core library.
main.cpp reduced to 15-line entry point.
No behavioral changes — pure file reorganization.
